refactor: remove throw on service
